requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-docx==0.8.11
lxml==4.9.3
psycopg2-binary==2.9.9
bcrypt==4.0.1
